grupo4/requirements.txt


#### requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
customtkinter>=5.0
Pillow>=9.0
python-dotenv>=0.21
joblib>=1.2
pytest>=7.0
